fix: 1. show session message if local storage exists 2. updated change log
git-svn-id: https://win-lhoa5a3q3i7/svn/front-end-backup/trunk@2622 3f25d8b1-e65f-8d48-bd54-75b34d8be7f1

+++ src/followups/default.js
@@ -34,7 +34,7 @@ module.exports = function (app) {
                 dialog
                     .errorMessage(langService.get('access_denied'))
             } else {
-                if (xhr.data.ec === 9002) {
+                if (xhr.data.ec === 9002 && localStorageService.get('CR')) {
                     localStorageService.remove('CR');
                     tokenService.destroy(); // destroy the current sessions
                     employeeService.destroyEmployee(); // destroy current user data
